requirements.txt
Flask
supabase
python-dotenv
elevenlabs
twilio
pyngrok